implementing reverse mode value -> colors
